requirements.txt
# requirements.txt

Flask==2.2.2
PyYAML==6.0
